backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
chromadb==0.5.0
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
pymupdf==1.24.3
google-generativeai==0.7.2
pydantic==2.7.1